+++ Tools/vscode/to2-syntax/client/src/extension.ts
@@ -12,7 +12,7 @@ let client: LanguageClient;
 export function activate(context: ExtensionContext) {
   // The server is implemented in node
   const serverModule = context.asAbsolutePath(
-    path.join("server", "out", "server.js")
+    path.join("server", "out", "lsp-server.js")
   );
 
   // If the extension is launched in debug mode then the debug server options are used