+++ Tools/vscode/to2-syntax/server/src/lsp-server.ts
@@ -1,187 +1,94 @@
 import { TextDocument } from "vscode-languageserver-textdocument";
-import { Registry } from "./to2/ast/registry";
 import {
   CompletionItem,
-  CompletionParams,
-  Connection,
-  Diagnostic,
-  DiagnosticSeverity,
-  DidChangeConfigurationParams,
-  Hover,
-  HoverParams,
-  SemanticTokens,
-  SemanticTokensParams,
-  TextDocumentChangeEvent,
-  TextDocuments,
-} from "vscode-languageserver";
-import { TextDocumentInput } from "./parser/text-document-input";
-import { module } from "./to2/parser-module";
-import { To2LspSettings, defaultSettings } from "./settings";
-import { TO2ModuleNode } from "./to2/ast/to2-module";
-import { SemanticToken, convertSemanticTokens } from "./syntax-token";
-import { findNodesAt } from "./helper";
+  CompletionItemKind,
+  DidChangeConfigurationNotification,
+  InitializeParams,
+  InitializeResult,
+  ProposedFeatures,
+  TextDocumentPositionParams,
+  TextDocumentSyncKind,
+  createConnection,
+} from "vscode-languageserver/node";
+import { LspServerSingleton } from "./lsp-server-singleton";
+import { SEMANTIC_TOKEN_MODIFIERS, SEMANTIC_TOKEN_TYPES } from "./syntax-token";
 
-export class LspServer {
-  private readonly registry = new Registry();
-  private globalSettings: To2LspSettings = defaultSettings;
-  private readonly documentSettings: Map<string, To2LspSettings> = new Map();
-  private readonly documentModules: Map<string, TO2ModuleNode> = new Map();
-  private readonly documents: TextDocuments<TextDocument>;
+const connection = createConnection(ProposedFeatures.all);
+const server = new LspServerSingleton(connection, false);
 
-  constructor(
-    private readonly connection: Connection,
-    public hasConfigurationCapability: boolean
-  ) {
-    this.documents = new TextDocuments(TextDocument);
-    this.documents.onDidOpen(this.onDidOpen.bind(this));
-    this.documents.onDidClose(this.onDidClose.bind(this));
-    this.documents.onDidChangeContent(this.onDidChange.bind(this));
-    this.documents.listen(connection);
-  }
+let hasWorkspaceFolderCapability = false;
+let hasDiagnosticRelatedInformationCapability = false;
 
-  async validateTextDocument(textDocument: TextDocument): Promise<void> {
-    const input = new TextDocumentInput(textDocument);
-    const moduleResult = module("<test>")(input);
+connection.onInitialize((params: InitializeParams) => {
+  const capabilities = params.capabilities;
 
-    const diagnostics: Diagnostic[] = [];
+  // Does the client support the `workspace/configuration` request?
+  // If not, we fall back using global settings.
+  server.hasConfigurationCapability = !!(
+    capabilities.workspace && !!capabilities.workspace.configuration
+  );
+  hasWorkspaceFolderCapability = !!(
+    capabilities.workspace && !!capabilities.workspace.workspaceFolders
+  );
+  hasDiagnosticRelatedInformationCapability = !!(
+    capabilities.textDocument &&
+    capabilities.textDocument.publishDiagnostics &&
+    capabilities.textDocument.publishDiagnostics.relatedInformation
+  );
 
-    if (moduleResult.value) {
-      this.documentModules.set(textDocument.uri, moduleResult.value);
-    } else {
-      this.documentModules.delete(textDocument.uri);
-    }
-
-    if (!moduleResult.success) {
-      const diagnostic: Diagnostic = {
-        severity: DiagnosticSeverity.Error,
-        range: {
-          start: moduleResult.remaining.position,
-          end: textDocument.positionAt(
-            moduleResult.remaining.position.offset +
-              moduleResult.remaining.available
-          ),
+  const result: InitializeResult = {
+    capabilities: {
+      textDocumentSync: TextDocumentSyncKind.Incremental,
+      // Tell the client that this server supports code completion.
+      completionProvider: {
+        resolveProvider: true,
+      },
+      semanticTokensProvider: {
+        legend: {
+          tokenTypes: [...SEMANTIC_TOKEN_TYPES],
+          tokenModifiers: [...SEMANTIC_TOKEN_MODIFIERS],
         },
-        message: moduleResult.expected,
-        source: "parser",
-      };
-      diagnostics.push(diagnostic);
-    }
-    if (moduleResult.value) {
-      // In this simple example we get the settings for every validate run.
-      const settings = await this.getDocumentSettings(textDocument.uri);
-
-      for (const validationError of moduleResult.value.validate(
-        this.registry
-      )) {
-        if (diagnostics.length < settings.maxNumberOfProblems) {
-          const diagnostic: Diagnostic = {
-            severity:
-              validationError.status === "error"
-                ? DiagnosticSeverity.Error
-                : DiagnosticSeverity.Warning,
-            range: validationError.range,
-            message: validationError.message,
-            source: "parser",
-          };
-          diagnostics.push(diagnostic);
-        }
-      }
-    }
-
-    // Send the computed diagnostics to VSCode.
-    this.connection.sendDiagnostics({ uri: textDocument.uri, diagnostics });
-  }
-
-  async getDocumentSettings(resource: string): Promise<To2LspSettings> {
-    if (!this.hasConfigurationCapability) {
-      return this.globalSettings;
-    }
-    let result = this.documentSettings.get(resource);
-    if (!result) {
-      result = await this.connection.workspace.getConfiguration({
-        scopeUri: resource,
-        section: "to2LspServer",
-      });
-      if (result) this.documentSettings.set(resource, result);
-    }
-    return result ?? this.globalSettings;
-  }
-
-  onDidChangeConfiguration(change: DidChangeConfigurationParams) {
-    if (this.hasConfigurationCapability) {
-      // Reset all cached document settings
-      this.documentSettings.clear();
-    } else {
-      this.globalSettings = <To2LspSettings>(
-        (change.settings.to2LspServer || defaultSettings)
-      );
-    }
-
-    // Revalidate all open text documents
-    this.documents
-      .all()
-      .forEach((document) => this.validateTextDocument(document));
-  }
-
-  onDidClose(event: TextDocumentChangeEvent<TextDocument>) {
-    this.documentSettings.delete(event.document.uri);
-    this.documentModules.delete(event.document.uri);
-  }
-
-  onDidOpen(event: TextDocumentChangeEvent<TextDocument>) {
-    this.validateTextDocument(event.document);
-  }
-
-  onDidChange(event: TextDocumentChangeEvent<TextDocument>) {
-    this.validateTextDocument(event.document);
-  }
-
-  onSemanticTokens(params: SemanticTokensParams): SemanticTokens {
-    const module = this.documentModules.get(params.textDocument.uri);
-    const token: SemanticToken[] = [];
-
-    module?.collectSemanticTokens(token);
-
-    return {
-      data: convertSemanticTokens(token),
+        full: true,
+      },
+      hoverProvider: true,
+      inlayHintProvider: true,
+    },
+  };
+  if (hasWorkspaceFolderCapability) {
+    result.capabilities.workspace = {
+      workspaceFolders: {
+        supported: true,
+      },
     };
   }
+  return result;
+});
 
-  onHover(params: HoverParams): Hover | undefined {
-    const module = this.documentModules.get(params.textDocument.uri);
-
-    if (!module) return undefined;
-
-    const documentation = findNodesAt(module, params.position).flatMap((node) =>
-      node.documentation
-        ? node.documentation
-            .filter((doc) => doc.range.contains(params.position))
-            .map((doc) => doc.value)
-        : []
-    );
-
-    if (documentation.length > 0)
-      return {
-        contents: {
-          kind: "markdown",
-          value: documentation.join("\n- - -\n"),
-        },
-      };
-
-    return undefined;
-  }
-
-  onCompleteion(params: CompletionParams): CompletionItem[] {
-    const module = this.documentModules.get(params.textDocument.uri);
-
-    if (!module) return [];
-
-    return findNodesAt(module, params.position).flatMap(
-      (node) => node.completionsAt?.(params.position) ?? []
+connection.onInitialized(() => {
+  if (server.hasConfigurationCapability) {
+    // Register for all configuration changes.
+    connection.client.register(
+      DidChangeConfigurationNotification.type,
+      undefined
     );
   }
-
-  onCompletionResolve(item: CompletionItem): CompletionItem {
-    return item;
+  if (hasWorkspaceFolderCapability) {
+    connection.workspace.onDidChangeWorkspaceFolders((_event) => {
+      connection.console.log("Workspace folder change event received.");
+    });
   }
-}
+});
+
+connection.onDidChangeWatchedFiles((_change) => {
+  // Monitored files have change in VSCode
+  connection.console.log("We received an file change event");
+});
+
+connection.onCompletion(server.onCompleteion.bind(server));
+connection.onCompletionResolve(server.onCompletionResolve.bind(server));
+connection.onHover(server.onHover.bind(server));
+connection.languages.inlayHint.on(server.onInlayHint.bind(server))
+connection.languages.semanticTokens.on(server.onSemanticTokens.bind(server));
+
+// Listen on the connection
+connection.listen();