+++ Tools/vscode/to2-syntax/server/src/to2/ast/to2-module.ts
@@ -1,3 +1,4 @@
+import { InlayHint } from "vscode-languageserver";
 import {
   ModuleItem,
   Node,