+++ Tools/vscode/to2-syntax/server/src/to2/ast/tuple-deconstruct-declaration.ts
@@ -7,10 +7,12 @@ import { TO2Type, UNKNOWN_TYPE } from "./to2-type";
 import { InputPosition, InputRange, WithPosition } from "../../parser";
 import { BlockContext } from "./context";
 import { SemanticToken } from "../../syntax-token";
+import { InlayHint } from "vscode-languageserver";
 
 export class TupleDeconstructDeclaration implements Node, BlockItem {
   public documentation?: WithPosition<string>[];
   public readonly range: InputRange;
+  public inlayHints: InlayHint[] | undefined;
 
   constructor(
     private readonly constLetKeyword: WithPosition<"let" | "const">,
@@ -38,6 +40,7 @@ export class TupleDeconstructDeclaration implements Node, BlockItem {
 
     const resultType = this.resultType(context).realizedType(context.module);
     this.documentation = [];
+    this.inlayHints = [];
     for (let i = 0; i < this.declarations.length; i++) {
       const declaration = this.declarations[i];
       if (isDeclarationParameter(declaration)) {
@@ -52,6 +55,13 @@ export class TupleDeconstructDeclaration implements Node, BlockItem {
             declaration.type?.value ??
             declaration.extractedType(resultType, i) ??
             UNKNOWN_TYPE;
+
+          if (variableType !== UNKNOWN_TYPE) {
+            this.inlayHints.push({
+              position: declaration.target.range.end,
+              label: ` : ${variableType.localName}`,
+            });
+          }
           context.localVariables.set(declaration.target.value, variableType);
           this.documentation.push(
             declaration.target.range.with(